requirements.txt
selenium
webdriver-manager
pandas
google-generativeai
requests
asyncio
openpyxl
playwright
aiohttp
Pillow
beautifulsoup4
pymongo
thefuzz